--- .vscode/settings.json
@@ -1,25 +1,0 @@
-{
-    // config related to code formatting
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "[javascript]": {
-        "editor.formatOnSave": false,
-        "editor.defaultFormatter": null
-    },
-    "[javascriptreact]": {
-        "editor.formatOnSave": false,
-        "editor.defaultFormatter": "esbenp.prettier-vscode"
-    },
-    "javascript.validate.enable": false, //disable all built-in syntax checking
-    "editor.codeActionsOnSave": {
-        "source.fixAll.eslint": true,
-        "source.fixAll.tslint": true,
-        "source.organizeImports": true
-    },
-    "eslint.alwaysShowStatus": true,
-    // emmet
-    "emmet.triggerExpansionOnTab": true,
-    "emmet.includeLanguages": {
-        "javascript": "javascriptreact"
-    }
-}
