--- client/.eslintrc.js
@@ -1,47 +1,0 @@
-{
-    "extends": [
-        "airbnb",
-        "airbnb/hooks",
-        "eslint:recommended",
-        "prettier",
-        "plugin:jsx-a11y/recommended"
-    ],
-    "parser": "babel-eslint",
-    "parserOptions": {
-        "ecmaVersion": 8
-    },
-    "env": {
-        "browser": true,
-        "node": true,
-        "es6": true,
-        "jest": true
-    },
-    "rules": {
-        "react/react-in-jsx-scope": 0,
-        "react-hooks/rules-of-hooks": "error",
-        "no-console": 0,
-        "react/state-in-constructor": 0,
-        "indent": 0,
-        "linebreak-style": 0,
-        "react/prop-types": 0,
-        "jsx-a11y/click-events-have-key-events": 0,
-        "react/jsx-filename-extension": [
-            1,
-            {
-                "extensions": [".js", ".jsx"]
-            }
-        ],
-        "prettier/prettier": [
-            "error",
-            {
-                "trailingComma": "es5",
-                "singleQuote": true,
-                "printWidth": 100,
-                "tabWidth": 4,
-                "semi": true,
-                "endOfLine": "auto"
-            }
-        ]
-    },
-    "plugins": ["prettier", "react", "react-hooks"]
-}
