+++ client/package.json
@@ -10,6 +10,7 @@
         "@testing-library/user-event": "^13.5.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
+        "react-infinite-scroll-component": "^6.1.0",
         "react-router-dom": "^6.8.0",
         "react-script": "^2.0.5",
         "react-scripts": "5.0.1",
