+++ client/src/components/Product/ProductSearch.jsx
@@ -1,0 +1,420 @@
+import { Dialog, Disclosure, Menu, Transition } from "@headlessui/react";
+import ProductInfinityScroll from "./ProductInfinityScroll";
+import {
+    ChevronDownIcon,
+    FunnelIcon,
+    MinusIcon,
+    PlusIcon,
+} from "@heroicons/react/20/solid";
+import { XMarkIcon } from "@heroicons/react/24/outline";
+import { Fragment, useState } from "react";
+
+const sortOptions = [
+    { name: "Most Popular", href: "#", current: true },
+    { name: "Best Rating", href: "#", current: false },
+    { name: "Newest", href: "#", current: false },
+    { name: "Price: Low to High", href: "#", current: false },
+    { name: "Price: High to Low", href: "#", current: false },
+];
+const subCategories = [
+    { name: "Totes", href: "#" },
+    { name: "Backpacks", href: "#" },
+    { name: "Travel Bags", href: "#" },
+    { name: "Hip Bags", href: "#" },
+    { name: "Laptop Sleeves", href: "#" },
+];
+const filters = [
+    {
+        id: "color",
+        name: "Color",
+        options: [
+            { value: "white", label: "White", checked: false },
+            { value: "beige", label: "Beige", checked: false },
+            { value: "blue", label: "Blue", checked: true },
+            { value: "brown", label: "Brown", checked: false },
+            { value: "green", label: "Green", checked: false },
+            { value: "purple", label: "Purple", checked: false },
+        ],
+    },
+    {
+        id: "category",
+        name: "Category",
+        options: [
+            { value: "new-arrivals", label: "New Arrivals", checked: false },
+            { value: "sale", label: "Sale", checked: false },
+            { value: "travel", label: "Travel", checked: true },
+            { value: "organization", label: "Organization", checked: false },
+            { value: "accessories", label: "Accessories", checked: false },
+        ],
+    },
+    {
+        id: "size",
+        name: "Size",
+        options: [
+            { value: "2l", label: "2L", checked: false },
+            { value: "6l", label: "6L", checked: false },
+            { value: "12l", label: "12L", checked: false },
+            { value: "18l", label: "18L", checked: false },
+            { value: "20l", label: "20L", checked: false },
+            { value: "40l", label: "40L", checked: true },
+        ],
+    },
+];
+
+function classNames(...classes) {
+    return classes.filter(Boolean).join(" ");
+}
+
+export default function ProductSearch({ products }) {
+    const [mobileFiltersOpen, setMobileFiltersOpen] = useState(false);
+    // a 12 item false array
+    const [selected, setSelected] = useState(new Array(12).fill(false));
+
+    return (
+        <div className="p-2">
+            {/* Mobile filter dialog */}
+            <Transition.Root show={mobileFiltersOpen} as={Fragment}>
+                <Dialog
+                    as="div"
+                    className="relative z-40 lg:hidden"
+                    onClose={setMobileFiltersOpen}
+                >
+                    <Transition.Child
+                        as={Fragment}
+                        enter="transition-opacity ease-linear duration-300"
+                        enterFrom="opacity-0"
+                        enterTo="opacity-100"
+                        leave="transition-opacity ease-linear duration-300"
+                        leaveFrom="opacity-100"
+                        leaveTo="opacity-0"
+                    >
+                        <div className="fixed inset-0 bg-black bg-opacity-25" />
+                    </Transition.Child>
+
+                    <div className="fixed inset-0 z-40 flex">
+                        <Transition.Child
+                            as={Fragment}
+                            enter="transition ease-in-out duration-300 transform"
+                            enterFrom="translate-x-full"
+                            enterTo="translate-x-0"
+                            leave="transition ease-in-out duration-300 transform"
+                            leaveFrom="translate-x-0"
+                            leaveTo="translate-x-full"
+                        >
+                            <Dialog.Panel className="relative ml-auto flex h-full w-full max-w-xs flex-col overflow-y-auto bg-white py-4 pb-12 shadow-xl">
+                                <div className="flex items-center justify-between px-4">
+                                    <h2 className="text-lg font-medium text-gray-900">
+                                        Filters
+                                    </h2>
+                                    <button
+                                        type="button"
+                                        className="-mr-2 flex h-10 w-10 items-center justify-center rounded-md bg-white p-2 text-gray-400"
+                                        onClick={() =>
+                                            setMobileFiltersOpen(false)
+                                        }
+                                    >
+                                        <span className="sr-only">
+                                            Close menu
+                                        </span>
+                                        <XMarkIcon
+                                            className="h-6 w-6"
+                                            aria-hidden="true"
+                                        />
+                                    </button>
+                                </div>
+
+                                {/* Filters */}
+                                <form className="mt-4 border-t border-gray-200">
+                                    <h3 className="sr-only">Categories</h3>
+                                    <ul
+                                        role="list"
+                                        className="px-2 py-3 font-medium text-gray-900"
+                                    >
+                                        {subCategories.map((category) => (
+                                            <li key={category.name}>
+                                                <a
+                                                    href={category.href}
+                                                    className="block px-2 py-3"
+                                                >
+                                                    {category.name}
+                                                </a>
+                                            </li>
+                                        ))}
+                                    </ul>
+
+                                    {filters.map((section) => (
+                                        <Disclosure
+                                            as="div"
+                                            key={section.id}
+                                            className="border-t border-gray-200 px-4 py-6"
+                                        >
+                                            {({ open }) => (
+                                                <>
+                                                    <h3 className="-mx-2 -my-3 flow-root">
+                                                        <Disclosure.Button className="flex w-full items-center justify-between bg-white px-2 py-3 text-gray-400 hover:text-gray-500">
+                                                            <span className="font-medium text-gray-900">
+                                                                {section.name}
+                                                            </span>
+                                                            <span className="ml-6 flex items-center">
+                                                                {open ? (
+                                                                    <MinusIcon
+                                                                        className="h-5 w-5"
+                                                                        aria-hidden="true"
+                                                                    />
+                                                                ) : (
+                                                                    <PlusIcon
+                                                                        className="h-5 w-5"
+                                                                        aria-hidden="true"
+                                                                    />
+                                                                )}
+                                                            </span>
+                                                        </Disclosure.Button>
+                                                    </h3>
+                                                    <Disclosure.Panel className="pt-6">
+                                                        <div className="space-y-6">
+                                                            {section.options.map(
+                                                                (
+                                                                    option,
+                                                                    optionIdx
+                                                                ) => (
+                                                                    <div
+                                                                        key={
+                                                                            option.value
+                                                                        }
+                                                                        className="flex items-center"
+                                                                    >
+                                                                        <input
+                                                                            id={`filter-mobile-${section.id}-${optionIdx}`}
+                                                                            name={`${section.id}[]`}
+                                                                            defaultValue={
+                                                                                option.value
+                                                                            }
+                                                                            type="checkbox"
+                                                                            defaultChecked={
+                                                                                option.checked
+                                                                            }
+                                                                            className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                        />
+                                                                        <label
+                                                                            htmlFor={`filter-mobile-${section.id}-${optionIdx}`}
+                                                                            className="ml-3 min-w-0 flex-1 text-gray-500"
+                                                                        >
+                                                                            {
+                                                                                option.label
+                                                                            }
+                                                                        </label>
+                                                                    </div>
+                                                                )
+                                                            )}
+                                                        </div>
+                                                    </Disclosure.Panel>
+                                                </>
+                                            )}
+                                        </Disclosure>
+                                    ))}
+                                </form>
+                            </Dialog.Panel>
+                        </Transition.Child>
+                    </div>
+                </Dialog>
+            </Transition.Root>
+
+            <main className="px-4">
+                <div className="flex items-baseline justify-between border-b border-gray-200">
+                    <label className="w-xl relative my-2 mr-2 flex-1 rounded-md lg:mr-8">
+                        <span class="absolute inset-y-0 left-0 flex items-center pl-2">
+                            <svg
+                                fill="bg-pink-500"
+                                viewBox="0 0 512 512"
+                                className="h-5 w-5 text-pink-100"
+                            >
+                                <path d="M479.6,399.716l-81.084-81.084-62.368-25.767A175.014,175.014,0,0,0,368,192c0-97.047-78.953-176-176-176S16,94.953,16,192,94.953,368,192,368a175.034,175.034,0,0,0,101.619-32.377l25.7,62.2L400.4,478.911a56,56,0,1,0,79.2-79.195ZM48,192c0-79.4,64.6-144,144-144s144,64.6,144,144S271.4,336,192,336,48,271.4,48,192ZM456.971,456.284a24.028,24.028,0,0,1-33.942,0l-76.572-76.572-23.894-57.835L380.4,345.771l76.573,76.572A24.028,24.028,0,0,1,456.971,456.284Z"></path>
+                            </svg>
+                        </span>
+                        <input
+                            className="block w-full rounded-md border border-slate-300 bg-white py-2 pl-9 pr-3 shadow-sm placeholder:italic placeholder:text-slate-400 focus:outline-none sm:text-sm"
+                            placeholder="Search for anything..."
+                            type="search"
+                            name="search"
+                        />
+                    </label>
+
+                    <div className="flex items-center">
+                        <Menu
+                            as="div"
+                            className="relative inline-block text-left"
+                        >
+                            <div>
+                                <Menu.Button className="group inline-flex justify-center text-sm font-medium text-gray-700 hover:text-gray-900">
+                                    Sort
+                                    <ChevronDownIcon
+                                        className="-mr-1 ml-1 h-5 w-5 flex-shrink-0 text-gray-400 group-hover:text-gray-500"
+                                        aria-hidden="true"
+                                    />
+                                </Menu.Button>
+                            </div>
+
+                            <Transition
+                                as={Fragment}
+                                enter="transition ease-out duration-100"
+                                enterFrom="transform opacity-0 scale-95"
+                                enterTo="transform opacity-100 scale-100"
+                                leave="transition ease-in duration-75"
+                                leaveFrom="transform opacity-100 scale-100"
+                                leaveTo="transform opacity-0 scale-95"
+                            >
+                                <Menu.Items className="absolute right-0 z-10 mt-2 w-40 origin-top-right rounded-md bg-white shadow-2xl ring-1 ring-black ring-opacity-5 focus:outline-none">
+                                    <div className="py-1">
+                                        {sortOptions.map((option) => (
+                                            <Menu.Item key={option.name}>
+                                                {({ active }) => (
+                                                    <a
+                                                        href={option.href}
+                                                        className={classNames(
+                                                            option.current
+                                                                ? "font-medium text-gray-900"
+                                                                : "text-gray-500",
+                                                            active
+                                                                ? "bg-gray-100"
+                                                                : "",
+                                                            "block px-4 py-2 text-sm"
+                                                        )}
+                                                    >
+                                                        {option.name}
+                                                    </a>
+                                                )}
+                                            </Menu.Item>
+                                        ))}
+                                    </div>
+                                </Menu.Items>
+                            </Transition>
+                        </Menu>
+
+                        {/* <button
+                            type="button"
+                            className="-m-2 ml-5 p-2 text-gray-400 hover:text-gray-500 sm:ml-7"
+                        >
+                            <span className="sr-only">View grid</span>
+                            <Squares2X2Icon
+                                className="h-5 w-5"
+                                aria-hidden="true"
+                            />
+                        </button> */}
+                        <button
+                            type="button"
+                            className="-m-2 ml-4 p-2 text-gray-400 hover:text-gray-500 sm:ml-6 lg:hidden"
+                            onClick={() => setMobileFiltersOpen(true)}
+                        >
+                            <span className="sr-only">Filters</span>
+                            <FunnelIcon
+                                className="h-5 w-5"
+                                aria-hidden="true"
+                            />
+                        </button>
+                    </div>
+                </div>
+
+                <section
+                    aria-labelledby="products-heading"
+                    className="pt-6 pb-24"
+                >
+                    <h2 id="products-heading" className="sr-only">
+                        Products
+                    </h2>
+
+                    <div className="grid grid-cols-1 gap-x-8 gap-y-10 lg:grid-cols-6">
+                        {/* Filters */}
+                        <form className="hidden lg:block">
+                            <h3 className="sr-only">Categories</h3>
+                            <ul
+                                role="list"
+                                className="space-y-4 border-b border-gray-200 pb-6 text-sm font-medium text-gray-900"
+                            >
+                                {subCategories.map((category) => (
+                                    <li key={category.name}>
+                                        <a href={category.href}>
+                                            {category.name}
+                                        </a>
+                                    </li>
+                                ))}
+                            </ul>
+
+                            {filters.map((section) => (
+                                <Disclosure
+                                    as="div"
+                                    key={section.id}
+                                    className="border-b border-gray-200 py-6"
+                                >
+                                    {({ open }) => (
+                                        <>
+                                            <h3 className="-my-3 flow-root">
+                                                <Disclosure.Button className="flex w-full items-center justify-between bg-white py-3 text-sm text-gray-400 hover:text-gray-500">
+                                                    <span className="font-medium text-gray-900">
+                                                        {section.name}
+                                                    </span>
+                                                    <span className="ml-6 flex items-center">
+                                                        {open ? (
+                                                            <MinusIcon
+                                                                className="h-5 w-5"
+                                                                aria-hidden="true"
+                                                            />
+                                                        ) : (
+                                                            <PlusIcon
+                                                                className="h-5 w-5"
+                                                                aria-hidden="true"
+                                                            />
+                                                        )}
+                                                    </span>
+                                                </Disclosure.Button>
+                                            </h3>
+                                            <Disclosure.Panel className="pt-6">
+                                                <div className="space-y-4">
+                                                    {section.options.map(
+                                                        (option, optionIdx) => (
+                                                            <div
+                                                                key={
+                                                                    option.value
+                                                                }
+                                                                className="flex items-center"
+                                                            >
+                                                                <input
+                                                                    id={`filter-${section.id}-${optionIdx}`}
+                                                                    name={`${section.id}[]`}
+                                                                    defaultValue={
+                                                                        option.value
+                                                                    }
+                                                                    type="checkbox"
+                                                                    defaultChecked={
+                                                                        option.checked
+                                                                    }
+                                                                    className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                />
+                                                                <label
+                                                                    htmlFor={`filter-${section.id}-${optionIdx}`}
+                                                                    className="ml-3 text-sm text-gray-600"
+                                                                >
+                                                                    {
+                                                                        option.label
+                                                                    }
+                                                                </label>
+                                                            </div>
+                                                        )
+                                                    )}
+                                                </div>
+                                            </Disclosure.Panel>
+                                        </>
+                                    )}
+                                </Disclosure>
+                            ))}
+                        </form>
+
+                        {/* Product grid */}
+                        <div className="lg:col-span-5">
+                            <ProductInfinityScroll products={products} />
+                        </div>
+                    </div>
+                </section>
+            </main>
+        </div>
+    );
+}
