+++ client/src/layout/NavBar/index.jsx
@@ -1,5 +1,6 @@
-import "./index.css";
 import { useState } from "react";
+import Logo from "../../assets/img/logo.svg";
+import "./index.css";
 
 export default function Index() {
     const [navbarOpen, setNavbarOpen] = useState(false);
@@ -30,20 +31,14 @@ export default function Index() {
     }
 
     return (
-        <nav className=" z-50 flex w-full flex-wrap items-center justify-between bg-gradient-to-r from-indigo-500 via-purple-500 to-pink-500 p-2">
+        <nav className=" z-50 flex w-full flex-wrap items-center justify-between bg-gradient-to-l from-indigo-500 via-purple-500 to-pink-500 p-2">
             <a
                 href="/"
                 className="mr-6 flex flex-shrink-0 items-center text-white"
             >
-                <svg
-                    className="mr-2 h-8 w-8 fill-current"
-                    width="54"
-                    height="54"
-                    viewBox="0 0 54 54"
-                    xmlns="http://www.w3.org/2000/svg"
-                >
-                    <path d="M13.5 22.1c1.8-7.2 6.3-10.8 13.5-10.8 10.8 0 12.15 8.1 17.55 9.45 3.6.9 6.75-.45 9.45-4.05-1.8 7.2-6.3 10.8-13.5 10.8-10.8 0-12.15-8.1-17.55-9.45-3.6-.9-6.75.45-9.45 4.05zM0 38.3c1.8-7.2 6.3-10.8 13.5-10.8 10.8 0 12.15 8.1 17.55 9.45 3.6.9 6.75-.45 9.45-4.05-1.8 7.2-6.3 10.8-13.5 10.8-10.8 0-12.15-8.1-17.55-9.45-3.6-.9-6.75.45-9.45 4.05z" />
-                </svg>
+                <div>
+                    <img src={Logo} alt="Think and Get Logo" className="w-10" />
+                </div>
                 <span className="text-xl font-semibold tracking-tight">
                     Think & Get
                 </span>
