--- jsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "ESNext",
-        "moduleResolution": "Node",
-        "target": "ES2020",
-        "jsx": "react",
-        "strictNullChecks": true,
-        "strictFunctionTypes": true
-    },
-    "exclude": ["node_modules", "**/node_modules/*"]
-}
